+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Fixed
 
 - Fixed Decompose Text node for fonts relative path.
+- Fixed fatal error when using replicate nodes.
 
 ### Changed